server/requirements.txt
openai
requests
Flask
python-dotenv
Flask-Scss
python-osc
